Round remaining time

+++ Samples/XboxGamingBarHelper/RTSS/OSDItems/OSDItemBattery.cs
@@ -41,8 +41,8 @@ namespace XboxGamingBarHelper.RTSS.OSDItems
 
                 if (batteryRemainTimeSensor.Value > 0)
                 {
-                    var hours = Math.Floor(batteryRemainTimeSensor.Value / MathConstants.SECONDS_PER_HOUR);
-                    var minutes = Math.Floor((batteryRemainTimeSensor.Value - hours * MathConstants.SECONDS_PER_HOUR) / MathConstants.SECONDS_PER_MINUTE);
+                    var hours = (int) Math.Floor(batteryRemainTimeSensor.Value / MathConstants.SECONDS_PER_HOUR);
+                    var minutes = (int) Math.Floor((batteryRemainTimeSensor.Value - hours * MathConstants.SECONDS_PER_HOUR) / MathConstants.SECONDS_PER_MINUTE);
                     osdItems.Add(new OSDItemValue(-1, $"{hours}<S=50> H<S> {minutes}<S=50> M<S>"));
                     //osdItems.Add(new OSDItemValue((float)hours, "H"));
                     //osdItems.Add(new OSDItemValue((float)minutes, "M"));
